Reorganized SceneGraph Nodes in a separate folder
