+++ src/brsTypes/index.ts
@@ -40,7 +40,7 @@ export * from "./components/RoList";
 export * from "./components/RoByteArray";
 export * from "./components/RoDateTime";
 export * from "./components/RoAssociativeArray";
-export * from "./components/Timespan";
+export * from "./components/RoTimespan";
 export * from "./components/BrsObjects";
 export * from "./components/RoRegex";
 export * from "./components/RoXMLElement";
@@ -53,23 +53,23 @@ export * from "./components/RoLongInteger";
 export * from "./components/RoInvalid";
 export * from "./components/RoSGNodeEvent";
 export * from "./components/RoSGNode";
-export * from "./components/Group";
-export * from "./components/Scene";
-export * from "./components/MiniKeyboard";
-export * from "./components/TextEditBox";
-export * from "./components/LayoutGroup";
-export * from "./components/Rectangle";
-export * from "./components/Label";
-export * from "./components/Font";
-export * from "./components/Poster";
-export * from "./components/ArrayGrid";
-export * from "./components/MarkupGrid";
-export * from "./components/ContentNode";
-export * from "./components/Timer";
+export * from "./nodes/Group";
+export * from "./nodes/Scene";
+export * from "./nodes/MiniKeyboard";
+export * from "./nodes/TextEditBox";
+export * from "./nodes/LayoutGroup";
+export * from "./nodes/Rectangle";
+export * from "./nodes/Label";
+export * from "./nodes/Font";
+export * from "./nodes/Poster";
+export * from "./nodes/ArrayGrid";
+export * from "./nodes/MarkupGrid";
+export * from "./nodes/ContentNode";
+export * from "./nodes/Timer";
 export * from "./components/RoAppInfo";
 export * from "./components/RoPath";
 export * from "./Boxing";
-export * from "./coercion";
+export * from "./Coercions";
 
 /**
  * Determines whether or not the given value is a number.
